+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optmiser/src/main/java/com/mmxlabs/scheduler/optimiser/components/IDischargeSlot.java
@@ -30,5 +30,5 @@ public interface IDischargeSlot extends IPortSlot {
 	 * 
 	 * @return
 	 */
-	long getSalesPrice();
+	int getSalesPrice();
 }
